requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
inflection==0.5.1
iniconfig==2.3.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
ruff==0.15.0
simplejwt==2.0.1
sqlparse==0.5.5
typing==3.7.4.3
tzdata==2025.3
uritemplate==4.2.0
django-storages==1.14.6
boto3==1.37.10
kubernetes==32.0.0
sentry-sdk==2.22.0
python-json-logger==2.0.7
firebase-admin==6.6.0
